GitHub workflow for a data-only master branch.
This workflow is intended to be run from the dev branch and pushes the generated xmltv files to the master branch.

The dev branch is now intended to be code only, while the master branch is data only without commit history.
This allows for normal code contribution and maintenance on the dev branch.